Sonar warnings and tests
