+++ packages/core/src/outcomes.test.ts
@@ -11,6 +11,7 @@ import {
   gone,
   isAccepted,
   isCreated,
+  isError,
   isGone,
   isNotFound,
   isOk,
@@ -122,6 +123,15 @@ describe('Outcomes', () => {
     expect(normalizeErrorString({ resourceType: 'OperationOutcome' })).toBe('Unknown error');
     expect(normalizeErrorString({ foo: 'bar' })).toBe('{"foo":"bar"}');
     expect(normalizeErrorString({ code: 'ERR_INVALID_ARG_TYPE' })).toBe('ERR_INVALID_ARG_TYPE');
+  });
+
+  test('isError', () => {
+    expect(isError(undefined)).toBe(false);
+    expect(isError(null)).toBe(false);
+    expect(isError('foo')).toBe(false);
+    expect(isError({ resourceType: 'Patient' })).toBe(false);
+    expect(isError(new Error('foo'))).toBe(true);
+    expect(isError(new DOMException('foo'))).toBe(true);
   });
 
   test('isOperationOutcome', () => {
